requirements.txt
streamlit
streamlit_option_menu
streamlit_tailwind
streamlit-aggrid    
streamlit_lottie
streamlit_extras
streamlit_shadcn_ui
streamlit-javascript
psycopg2-binary  
supabase         
python-dotenv
pandas
numpy
altair
Pillow
pdf2image
fpdf
python-docx
xlsxwriter
httpx
datetime
google-generativeai
plotly